Remove eunit test fun - don't call this test
Rather than remove the WIP file, instead remove the eunit test function
so that the statem is not called automatically as part of the test
proceedure. Though this statem has been of use in fixing the claim
issues, and verifying the fixes, it is not yet ready. Still to be done
is the work that actually removes a node and it's ownership from the ring.